refactor: ProductCreateViewController 리팩토링 #8
