--- OpenMarket/OpenMarket/Controller/ProductModelManager.swift
@@ -1,68 +1,0 @@
-//
-//  ProductModelManager.swift
-//  OpenMarket
-//
-//  Created by JeongTaek Han on 2022/01/26.
-//
-
-import Foundation
-import UIKit.UIImage
-
-class ProductModelManager {
-    
-    private let networkManager = ProductNetworkManager()
-    
-    private let id: Int
-    private(set) var product: Product?
-    private(set) var images: [UIImage] = []
-    
-    let updateHandler: (() -> Void)
-    
-    init(id: Int, modelHandler: @escaping (() -> Void)) {
-        self.id = id
-        self.updateHandler = modelHandler
-    }
-    
-    func requestDetailProduct() {
-        networkManager.detailProductRequest(id: id) { result in
-            switch result {
-            case .success(let data):
-                self.product = data
-                self.requestProductImages()
-                self.updateHandler()
-            case .failure(let error):
-                print(error)
-            }
-        }
-    }
-    
-    func requestModifyProduct(form: ProductRegisterForm,
-                              completionHandler: ((Result<Product, URLSessionProviderError>) -> Void)? = nil) {
-        let requestForm = UpdateProductRequestModel(
-            name: form.name,
-            descriptions: form.description,
-            thumbnailID: nil,
-            price: Double(form.price),
-            currency: Currency(rawValue: form.currency),
-            discountedPrice: Double(form.discountedPrice ?? ""),
-            secret: AppConfigure.venderSecret
-        )
-        networkManager.modifyProductRequest(id: id, requestForm: requestForm, completionHandeler: completionHandler)
-    }
-    
-    private func requestProductImages() {
-        product?.images?.forEach({ productImage in
-            let url = productImage.url
-            networkManager.fetchImages(url: url) { result in
-                switch result {
-                case.success(let image):
-                    self.images.append(image)
-                    self.updateHandler()
-                case .failure(let error):
-                    print(error)
-                }
-            }
-        })
-    }
-    
-}
